+++ src/main/java/com/minecolonies/coremod/colony/jobs/JobBlacksmith.java
@@ -1,0 +1,41 @@
+package com.minecolonies.coremod.colony.jobs;
+
+import com.minecolonies.coremod.colony.CitizenData;
+import com.minecolonies.coremod.entity.ai.basic.AbstractAISkeleton;
+import com.minecolonies.coremod.entity.ai.citizen.blacksmith.EntityAIWorkBlacksmith;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Class of the Blacksmith job.
+ */
+public class JobBlacksmith extends AbstractJobCrafter
+{
+    /**
+     * Instantiates the job for the Blacksmith.
+     *
+     * @param entity the citizen who becomes a Sawmill
+     */
+    public JobBlacksmith(final CitizenData entity)
+    {
+        super(entity);
+    }
+
+    @NotNull
+    @Override
+    public String getName()
+    {
+        return "com.minecolonies.coremod.job.Sawmill";
+    }
+
+    /**
+     * Generate your AI class to register.
+     *
+     * @return your personal AI instance.
+     */
+    @NotNull
+    @Override
+    public AbstractAISkeleton<JobBlacksmith> generateAI()
+    {
+        return new EntityAIWorkBlacksmith(this);
+    }
+}
