+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/blacksmith/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all the classes the blacksmith uses to work.
+ */
+package com.minecolonies.coremod.entity.ai.citizen.blacksmith;
