+++ src/main/resources/assets/minecolonies/lang/en_us.lang
@@ -4,7 +4,7 @@ tile.minecolonies.blockHutSwineHerder.name=SwineHerders's Hut
 tile.minecolonies.blockHutCowboy.name=Cowboy's Hut
 tile.minecolonies.blockHutChickenHerder.name=ChickenHerders's Hut
 tile.minecolonies.blockHutBaker.name=Bakery
-tile.minecolonies.blockHutBlacksmith.name=Forge (NOT IMPLEMENTED)
+tile.minecolonies.blockHutBlacksmith.name=Blacksmith
 tile.minecolonies.blockHutBuilder.name=Builder's Hut
 tile.minecolonies.blockHutFarmer.name=Farm
 tile.minecolonies.blockHutLumberjack.name=Lumberjack's Hut
